RDKAAMP-2063 [l2test] generate a test report compatible with RDK-E requirements
Reason for change:
Convert test to use pytest framework which gives an easy route to .json test results

Test Procedure: Run l2 tests at aamp/test/l2test.

Risks: Low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_24_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_24_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_24_1

+++ test/l2test/TST_2000_Simlinear/prerequisite.sh
@@ -8,6 +8,12 @@ if [ "$TEST_2000_STREAM_PATH" == "" ]; then
     TEST_2000_STREAM_PATH="https://cpetestutility.stb.r53.xcal.tv/AAMP/simlinear/aamptest/streams/simlinear/streams09-Jan2023.tar.gz"
 fi
 
+#Stop multiple fetching of data if we already have it
+file=$(basename ${TEST_2000_STREAM_PATH})
+if [ -f $file ]; then
+    echo "$file exists, not fetching again"
+    exit 0
+fi
 if (curl -O  -X GET $TEST_2000_STREAM_PATH);then
     echo "Test referring :" $TEST_2000_STREAM_PATH
 else
